+++ Demo/Demo/AppDelegate.m
@@ -1,5 +1,5 @@
 #import "AppDelegate.h"
-#import <PopupBridge/POPPopupBridge.h>
+#import <PopupBridge/PopupBridge-Swift.h>
 
 @interface AppDelegate ()