+++ Demo/Demo/POPViewController.m
@@ -1,6 +1,6 @@
 #import "POPViewController.h"
 #import <WebKit/WebKit.h>
-#import <PopupBridge/POPPopupBridge.h>
+#import <PopupBridge/PopupBridge-Swift.h>
 
 @interface POPViewController () <POPPopupBridgeDelegate>
 @property (nonatomic, strong) WKWebView *webView;