+++ Demo/Demo/SceneDelegate.m
@@ -1,5 +1,5 @@
 #import "SceneDelegate.h"
-#import <PopupBridge/POPPopupBridge.h>
+#import <PopupBridge/PopupBridge-Swift.h>
 
 @interface SceneDelegate ()
 
@@ -11,7 +11,7 @@
    for (UIOpenURLContext *urlContext in URLContexts) {
        NSURL *url = [urlContext URL];
        if ([url.scheme localizedCaseInsensitiveCompare:@"com.braintreepayments.popupbridgeexample"] == NSOrderedSame) {
-       [POPPopupBridge openURL:urlContext.URL];
+           [POPPopupBridge openURL:urlContext.URL];
        }
    }
 }