+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:5.3
+// swift-tools-version:5.7
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 import PackageDescription
 
@@ -14,8 +14,7 @@ let package = Package(
     targets: [
         .target(
             name: "PopupBridge",
-            exclude: ["PopupBridge-Framework-Info.plist"],
-            publicHeadersPath: "Public"
+            exclude: ["PopupBridge-Framework-Info.plist"]
         ),
     ]
 )