+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # PopupBridge iOS Release Notes
 
+## unreleased
+
+* Convert PopupBridge to Swift
+* Breaking Changes
+  * Remove deprecated `POPPopupBridge.open(url:sourceApplication:)` & `POPPopupBridge.open(url:options:)` methods
+  * Bump Package.swift and PopupBridge.podspec to Swift version 5.7
+
 ## 1.2.0 (2021-01-22)
 
 * Exclude arm64 simulator architectures via Podspec